Move key classes to dedicated SPI
