+++ extensions/common/iam/oauth2/oauth2-core/src/main/java/org/eclipse/edc/iam/oauth2/identity/IdentityProviderKeyResolver.java
@@ -18,9 +18,9 @@ package org.eclipse.edc.iam.oauth2.identity;
 import okhttp3.Request;
 import org.eclipse.edc.iam.oauth2.jwt.JwkKey;
 import org.eclipse.edc.iam.oauth2.jwt.JwkKeys;
+import org.eclipse.edc.keys.spi.PublicKeyResolver;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.http.EdcHttpClient;
-import org.eclipse.edc.spi.iam.PublicKeyResolver;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.result.Result;
 import org.eclipse.edc.spi.types.TypeManager;
