+++ extensions/control-plane/transfer/transfer-data-plane/src/test/java/org/eclipse/edc/connector/transfer/dataplane/api/ConsumerPullTransferTokenValidationApiControllerTest.java
@@ -15,8 +15,8 @@
 package org.eclipse.edc.connector.transfer.dataplane.api;
 
 import org.eclipse.edc.connector.transfer.dataplane.spi.security.DataEncrypter;
+import org.eclipse.edc.keys.spi.PublicKeyResolver;
 import org.eclipse.edc.spi.iam.ClaimToken;
-import org.eclipse.edc.spi.iam.PublicKeyResolver;
 import org.eclipse.edc.spi.result.Result;
 import org.eclipse.edc.spi.types.TypeManager;
 import org.eclipse.edc.spi.types.domain.DataAddress;
