+++ spi/common/token-spi/src/main/java/org/eclipse/edc/token/spi/TokenValidationService.java
@@ -14,8 +14,8 @@
 
 package org.eclipse.edc.token.spi;
 
+import org.eclipse.edc.keys.spi.PublicKeyResolver;
 import org.eclipse.edc.spi.iam.ClaimToken;
-import org.eclipse.edc.spi.iam.PublicKeyResolver;
 import org.eclipse.edc.spi.iam.TokenRepresentation;
 import org.eclipse.edc.spi.result.Result;
 import org.jetbrains.annotations.NotNull;
